+++ src/main.cpp
@@ -74,12 +74,12 @@ int main(int argc, char* argv[]) {
 
   MomoArgs args;
 
-  bool use_test = false;
+  bool use_p2p = false;
   bool use_ayame = false;
   bool use_sora = false;
   int log_level = webrtc::LS_NONE;
 
-  Util::ParseArgs(argc, argv, use_test, use_ayame, use_sora, log_level, args);
+  Util::ParseArgs(argc, argv, use_p2p, use_ayame, use_sora, log_level, args);
 
   webrtc::LogMessage::LogToDebug((webrtc::LoggingSeverity)log_level);
   webrtc::LogMessage::LogTimestamps();
@@ -343,14 +343,14 @@ int main(int argc, char* argv[]) {
       stats_collector = sora_client;
     }
 
-    if (use_test) {
+    if (use_p2p) {
       P2PServerConfig config;
       config.no_google_stun = args.no_google_stun;
-      config.doc_root = args.test_document_root;
+      config.doc_root = args.p2p_document_root;
 
       const boost::asio::ip::tcp::endpoint endpoint{
           boost::asio::ip::make_address("0.0.0.0"),
-          static_cast<unsigned short>(args.test_port)};
+          static_cast<unsigned short>(args.p2p_port)};
       p2p_server = P2PServer::Create(ioc, endpoint, rtc_manager.get(),
                                      std::move(config));
       p2p_server->Run();